+++ src/ReadoutVersion.h
@@ -1,0 +1,1 @@
+#define READOUT_VERSION "1.3.9.x"